Add support for searching and displaying Pokemon types.
The `SearchBar` component now imports the `pokemonTypes` array from the `utils` module. The `matchingType` array has been added to store the matching types. The `typeData` array has been initialized with the `pokemonTypes` array.

Inside the `updateMatching` function, a loop has been added to search for matching types based on the search term. The matched types are stored in the `matchingType` array.

In the template, a new `{#each}` block has been added to display the matching types. Each type is rendered as a clickable link with an associated `Type` component.

In the `Type` component, the wrapping `<div>` has been replaced with an `<a>` tag with a `data-sveltekit-reload` attribute for routing. The `<div