feat: Add mobile support for search drawer
This commit introduces mobile support for the search drawer. Now, when the search icon is clicked on mobile devices, the drawer will open. This feature is implemented by toggling the display of the drawer component when the button is clicked. The layout of the search results on mobile devices has also been updated to display the results in a single column, instead of the previous three columns on larger screens.

The commit also includes an additional code change in the `pokemon/[slug]` page. The order of the rendering elements has been rearranged to provide a better user experience. Specifically, the Pokemon image is now displayed above the Pokemon name, and the styling of the name is updated to dynamically use a different color for legendary Pokemon based on the data from the API response.